refactor: estructuring the project in folders
acabo de re-estructurar todo el proyecto para tenerlo todo mas
organizado en carpetas ya actualice tambien los enlaces vinculados en
cada documento pero puede que haya quedado alguno por actualizar, se ira
viendo con el avance
