+++ app/App.js
@@ -1,5 +1,5 @@
 import { Router } from "./components/Router.js";
-import { ScrollTopButton } from "./components/ScrollTopButt.js";
+import { ScrollTopButton } from "./helpers/ScrollTopButt.js";
 
 export default function App() {
   Router();
