+++ app/assets/Loader.js
