+++ app/components/singers/CardSing.js
