+++ app/components/singers/S_ComInput.js
